Fix crash on policy ls
Policy ls would crash due to a nil pointer dereference when there were
no policies on the server. This was due to the formatting behaviour in
the pretty printer not playing well with the data type provided with the
go-api.

The fix has been made inside the go-api repo, this commit updates the
vendored copy to include this fix.