DEV-1699: CLI login accepts root -u & -p flags
Will first try and use subcommand -u & -p flags, then root -u & -p flags
and then if those aren't present will try and use environment variables.
If one is present but not the other (at any level) then the missing part
of the credentials will be requested via a prompt. Precedence is
right-to-left.