Fix wrong fstype mount
Volumes could be initialised and mounted with a different fstype value
from those specified at volume create time. This is now not possible as
mkfs and mount take their fs type value from the controlplane volume object.

Adds a fstype object which provides a standard set of methods for
generating mkfs and mount commands. This centralises this piece of
logic, hopefully making bugs easier to spot.