mount: use node's DeviceDir not the default path
This changes how we determine if we are on a storageos node. Using the
hostname, the client checks if there exists any registered node with the
current hostname. If not, mounting fails with appropriate error.

Once we obtain a node object, we use the node.DeviceDir to check if
StorageOS device dir exists in that node. The same is used in
retryableMount.