Humanise hostname ordering
Many commands list the hostname as the first column of output but order
the list lexicographically. This causes hostnames with trailing numeric
identifiers (e.g. host{1..11}) to appear in a unexpected order
(1,11,2,3,...).

This commit attempts to address this by trimming shared sub strings
before sorting, and attempting to parse the remainder as an int where
possible.